scenario13-requirements.txt
opentelemetry-sdk==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0